Make float database operations more consistent.
We don't use floats in GnuCash, we use doubles (and those as little as
possible), but dbd-sqlite3 is broken in that it stores only floats.
Simply casting floats to doubles introduces bogus additional digits
that can cause round-trip tests to fail. Instead convert floats to
doubles by multiplying by 10E6, rounding, then dividing by 10E6.

+++ libgnucash/backend/sql/gnc-sql-column-table-entry.cpp
@@ -283,7 +283,7 @@ GncSqlColumnTableEntryImpl<CT_DOUBLE>::load (const GncSqlBackend* sql_be,
     {
         try
         {
-            val = static_cast<double>(row.get_float_at_col(m_col_name));
+            val = row.get_float_at_col(m_col_name);
         }
         catch (std::invalid_argument&)
         {
